requirements.txt
adal==1.2.7
applicationinsights==0.11.10
azure-common==1.1.27
azure-core==1.19.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.1.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.1.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-core==1.34.0
azureml-datadrift==1.34.0
azureml-dataprep==2.22.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==1.20.2
azureml-dataset-runtime==1.34.0
azureml-defaults==1.34.0
azureml-inference-server-http==0.3.2
azureml-opendatasets==1.34.0
azureml-pipeline-core==1.34.0
azureml-telemetry==1.34.0
backports.tempfile==1.0
backports.weakref==1.0.post1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cloudpickle==1.6.0
configparser==3.7.4
contextlib2==21.6.0
cryptography==3.4.8
cycler==0.10.0
distro==1.6.0
docker==5.0.2
dotnetcore2==2.1.21
Flask==1.0.3
fusepy==3.0.1
google-api-core==1.31.3
google-auth==1.35.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
inference-schema==1.3.0
isodate==0.6.0
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json-logging-py==0.2
jsonpickle==2.0.0
kiwisolver==1.3.2
lightgbm==3.2.1
MarkupSafe==2.0.1
matplotlib==3.2.1
msal==1.14.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.21.2
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
packaging==21.0
pandas==1.3.3
pathspec==0.9.0
portalocker==1.7.1
protobuf==3.17.3
psutil==5.8.0
py4j==0.10.9
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyspark==3.1.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
scikit-learn==1.0
scipy==1.7.1
SecretStorage==3.3.1
six==1.16.0
threadpoolctl==2.2.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.12.1
